# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food Booking API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bookings
    description: Booking operations
  - name: menu
    description: Menu item operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Pending, Confirmed, Cancelled, Completed]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /bookings/stats:
    get:
      operationId: getBookingStats
      tags: [bookings]
      summary: Get booking statistics
      responses:
        "200":
          description: Booking stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingStats"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: availableOnly
          in: query
          required: false
          schema:
            type: boolean
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BookingStatus:
      type: string
      enum: [Pending, Confirmed, Cancelled, Completed]

    Booking:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
        guests:
          type: integer
        mealType:
          type: string
        specialRequests:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookingStatus"
        bookedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - email
        - date
        - time
        - guests
        - mealType
        - status
        - bookedAt

    CreateBookingBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        date:
          type: string
          format: date
        time:
          type: string
          minLength: 1
        guests:
          type: integer
          minimum: 1
          maximum: 20
        mealType:
          type: string
          minLength: 1
        specialRequests:
          type: string
      required:
        - name
        - phone
        - email
        - date
        - time
        - guests
        - mealType

    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/BookingStatus"
      required:
        - status

    BookingStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        confirmed:
          type: integer
        cancelled:
          type: integer
        completed:
          type: integer
        todayCount:
          type: integer
      required:
        - total
        - pending
        - confirmed
        - cancelled
        - completed
        - todayCount

    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        pricePaise:
          type: integer
          description: Price in paise (1/100 of INR). e.g. 25000 = ₹250.00
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        available:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - pricePaise
        - category
        - available
        - sortOrder
        - createdAt
        - updatedAt

    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        pricePaise:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
        imageUrl:
          type: string
        available:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0
      required:
        - name
        - pricePaise
        - category

    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        pricePaise:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
        imageUrl:
          type: string
          nullable: true
        available:
          type: boolean
        sortOrder:
          type: integer
          minimum: 0

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error